requirements.txt
Flask==3.0.3
torch==1.11.0
torchvision==0.12.0
Pillow==9.1.0
opencv-python==4.5.5.64
streamlit==1.36.0